Refactor: 스마트맵, 상품 추천 페이지 리팩토링
